Consolidate all the promo label components
Reduce the duplication of promo label component, tests and translation files
