--- blocks/extra-large-promo-block/intl.json
@@ -1,22 +1,0 @@
-{
-   "extra-large-promo-block.gallery-text":{
-      "en":"Gallery",
-      "sv":"Bildspel",
-      "no":"Bildegalleri",
-      "fr": "Galerie",
-      "de": "Galerie",
-      "es": "Galería",
-      "ja": "ギャラリー",
-      "ko": "갤러리"
-   },
-   "extra-large-promo-block.video-text":{
-      "en":"Video",
-      "sv":"Video",
-      "no":"Video",
-      "fr": "Vidéo",
-      "de": "Video",
-      "es": "Video",
-      "ja": "ビデオ",
-      "ko": "비디오"
-   }
-}
