+++ blocks/shared-styles/index.js
@@ -1,3 +1,4 @@
+import PromoLabel from './_children/promo-label';
 import PrimaryFont from './_children/primary-font';
 import SecondaryFont from './_children/secondary-font';
 import VSpace from './_children/v-space/index';
@@ -5,6 +6,7 @@ import { breakpointCustomField, spacerCustomField } from './_children/v-space/cu
 
 export {
   breakpointCustomField,
+  PromoLabel,
   PrimaryFont,
   SecondaryFont,
   spacerCustomField,
