+++ blocks/shared-styles/intl.json
@@ -1,0 +1,22 @@
+{
+  "promo-label.gallery-text":{
+    "en":"Gallery",
+    "sv":"Bildspel",
+    "no":"Bildegalleri",
+    "fr": "Galerie",
+    "de": "Galerie",
+    "es": "Galería",
+    "ja": "ギャラリー",
+    "ko": "갤러리"
+  },
+  "promo-label.video-text":{
+    "en":"Video",
+    "sv":"Video",
+    "no":"Video",
+    "fr": "Vidéo",
+    "de": "Video",
+    "es": "Video",
+    "ja": "ビデオ",
+    "ko": "비디오"
+  }
+}
