+++ blocks/top-table-list-block/features/top-table-list/_children/small-list-item.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Image } from '@wpmedia/engine-theme-sdk';
 import { ratiosFor } from '@wpmedia/resizer-image-block';
 import getProperties from 'fusion:properties';
+import { PromoLabel } from '@wpmedia/shared-styles';
 import Title from './title';
-import PromoLabel from './promo_label';
 import discoverPromoType from './discover';
 import {
   LEFT, RIGHT, ABOVE, BELOW,
