+++ blocks/top-table-list-block/features/top-table-list/_children/vertical-overline-image-story-item.jsx
@@ -9,10 +9,10 @@ import Byline from '@wpmedia/byline-block';
 import Overline from '@wpmedia/overline-block';
 import { ratiosFor } from '@wpmedia/resizer-image-block';
 import getProperties from 'fusion:properties';
+import { PromoLabel } from '@wpmedia/shared-styles';
 import Title from './title';
 import DescriptionText from './description-text';
 import checkObjectEmpty from '../shared/checkObjectEmpty';
-import PromoLabel from './promo_label';
 import discoverPromoType from './discover';
 
 const VerticalOverlineImageStoryItem = (props) => {
