--- blocks/top-table-list-block/intl.json
@@ -1,14 +1,0 @@
-{
-   "top-table-list-block.gallery-text":{
-      "en":"Gallery",
-      "sv":"Bildspel",
-      "no":"Bildegalleri",
-      "fr": "Galerie"
-   },
-   "top-table-list-block.video-text":{
-      "en":"Video",
-      "sv":"Video",
-      "no":"Video",
-      "fr": "Vidéo"
-   }
-}
